Update README
The commit refactors the service architecture in the `README.md` file, changing the section title from "Design" to "Service Architecture". It also adds low-level details about the orderbook implementation, including code snippets and explanations of data structures used. The commit aims to provide a better understanding of how orders are stored and accessed in the orderbook.