feat: Add Redis configuration options
This commit adds support for configuring the Redis host and port in the application. The `Config` struct now includes fields for `redis_url`, `redis_host`, and `redis_port`. The default values are set to "redis" for the host and 6379 for the port. If environment variables `REDIS_HOST` and/or `REDIS_PORT` are provided, they will be used instead of the defaults.